fix false packet loss from ICMP burst rate-limits
Keep the same per-target interval/count (measurement precision unchanged)
but pace full-mesh sends round-robin and stagger node start phases so
agents no longer fire N targets in lockstep. Raise ICMP reply quotas in
deploy/sysctl, and store loss percent with two decimal places instead of
truncating sub-1% samples to zero.

+++ src/g/struct.go
@@ -59,7 +59,8 @@ func (n *NetworkMember) UnmarshalJSON(b []byte) error {
 type PingSt struct {
 	SendPk   int
 	RevcPk   int
-	LossPk   int
+	// LossPk 丢包率百分比(可带小数, 如 0.37); 保留精度避免 int 截断低估
+	LossPk   float64
 	MinDelay float64
 	AvgDelay float64
 	MaxDelay float64
